+++ srcs/tools3.c
@@ -6,7 +6,7 @@
 /*   By: jjaegle <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/19 16:19:53 by jjaegle           #+#    #+#             */
-/*   Updated: 2019/11/28 15:43:28 by jjaegle          ###   ########.fr       */
+/*   Updated: 2019/11/29 12:02:43 by jjaegle          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ void	set_param_value(unsigned char *mem, t_process *process, int i
 	}
 	else if (process->param[i - 1].type == REG_CODE)
 		process->param[i - 1].value =
-			*(int*)process->records[(short)process->param[i - 1].value - 1];
+			*(int*)process->records[process->param[i - 1].value - 1];
 }
 
 void	write_in_visu(int start, int dest, t_env *vm)
@@ -63,8 +63,9 @@ void	del_visu(t_visu *visu)
 	{
 		attron(COLOR_PAIR(1));
 		if (visu->vm->lastlive != -1)
-			mvprintw(LINES / 2 - 3, (COLS / 2) - 10, "Le joueur %d (%s) a gagné",
-			visu->vm->lastlive, visu->vm->tab_champ.champs[visu->vm->lastlive].name);
+			mvprintw(LINES / 2 - 3, (COLS / 2) - 10
+					, "Le joueur %d (%s) a gagné", visu->vm->lastlive
+					, visu->vm->tab_champ.champs[visu->vm->lastlive].name);
 		else
 			mvprintw(LINES / 2 - 3, (COLS / 2) - 10
 					, "No winner for today... Only looser");
